# Journal-front/src/app/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8083",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/{idman}/comments/save": {
      "post": {
        "tags": [
          "comment"
        ],
        "operationId": "saveComment",
        "parameters": [
          {
            "name": "idman",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDTO"
                }
              }
            }
          }
        }
      }
    },
    "/{idauth}/manuscripts/save": {
      "post": {
        "tags": [
          "manuscript"
        ],
        "operationId": "saveManuscript",
        "parameters": [
          {
            "name": "idauth",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManuscriptDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManuscriptDTO"
                }
              }
            }
          }
        }
      }
    },
    "/{idEdit}/invitations/save/{idRev}/{idMan}": {
      "post": {
        "tags": [
          "invitation"
        ],
        "operationId": "sendInvitation",
        "parameters": [
          {
            "name": "idEdit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idRev",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "idMan",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/manuscripts/reject/{id}": {
      "post": {
        "tags": [
          "manuscript"
        ],
        "operationId": "RejectManuscriptById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/{idman}/comments": {
      "get": {
        "tags": [
          "comment"
        ],
        "operationId": "getComments",
        "parameters": [
          {
            "name": "idman",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{idauth}/manuscripts": {
      "get": {
        "tags": [
          "manuscript"
        ],
        "operationId": "getManuscripts",
        "parameters": [
          {
            "name": "idauth",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManuscriptDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{idauth}/manuscripts/waiting": {
      "get": {
        "tags": [
          "manuscript"
        ],
        "operationId": "getWaitingManuscripts",
        "parameters": [
          {
            "name": "idauth",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManuscriptDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{idauth}/manuscripts/sentBack": {
      "get": {
        "tags": [
          "manuscript"
        ],
        "operationId": "getSentBackManuscripts",
        "parameters": [
          {
            "name": "idauth",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManuscriptDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{idauth}/manuscripts/processed": {
      "get": {
        "tags": [
          "manuscript"
        ],
        "operationId": "getProcessedManuscripts",
        "parameters": [
          {
            "name": "idauth",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManuscriptDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{idauth}/manuscripts/incomplete": {
      "get": {
        "tags": [
          "manuscript"
        ],
        "operationId": "getIncompleteManuscripts",
        "parameters": [
          {
            "name": "idauth",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManuscriptDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/{idRev}/invitations": {
      "get": {
        "tags": [
          "invitation"
        ],
        "operationId": "getInvitationsByReviewer",
        "parameters": [
          {
            "name": "idRev",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/{idEdit}/invitations": {
      "get": {
        "tags": [
          "invitation"
        ],
        "operationId": "getInvitationsByEditor",
        "parameters": [
          {
            "name": "idEdit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/manuscripts/{idman}/authors/add/{author}": {
      "get": {
        "tags": [
          "manuscript"
        ],
        "operationId": "addAuthor",
        "parameters": [
          {
            "name": "idman",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "author",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuthorDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invitations/{id}": {
      "get": {
        "tags": [
          "invitation"
        ],
        "operationId": "getInvitationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/manuscripts/delete/{id}": {
      "delete": {
        "tags": [
          "manuscript"
        ],
        "operationId": "deleteManuscriptById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/invitations/delete/{id}": {
      "delete": {
        "tags": [
          "invitation"
        ],
        "operationId": "deleteInvitationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/comments/delete/{id}": {
      "delete": {
        "tags": [
          "comment"
        ],
        "operationId": "deleteCommentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "ManuscriptDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          }
        }
      },
      "EditorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "InvitationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "editorDTO": {
            "$ref": "#/components/schemas/EditorDTO"
          },
          "reviewerDTO": {
            "$ref": "#/components/schemas/ReviewerDTO"
          },
          "manuscriptDTO": {
            "$ref": "#/components/schemas/ManuscriptDTO"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReviewerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "AuthorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "jobTitle": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    }
  }
}